requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
pydantic>=2.6
pydantic-settings>=2.2
argon2-cffi>=23.1
email-validator>=2.1
